+++ test/smooth.jl
@@ -1,30 +1,28 @@
 using Spectra
 using Test
+using Statistics
+using Random
 
 # based on the example of the smooth function...
 
 # the x axis
-x = collect(0:0.1:100)
-
-# a scale factor
-scale = 0.01
+x = collect(0:0.1:10)
 
 # a fake signal: perfect y
-y_tot, y_peaks = gaussiennes([10;20.],[40.;60],[5.;15],x)
-y_perfect = scale.*y_tot
+y_perfect = x.^2
 
-# we add noise: observed y
-y = scale.*(y_tot + randn(size(y_tot,1)))
+# fake noisy signal
+y = y_perfect + randn(size(x,1))*2
 
-y_sv = smooth(x,y,filter=:SavitzkyGolay,M=15,N=2)
-y_gcv = smooth(x,y,filter=:GCVSmoothedNSpline, ese_y = std(y))
-y_whit = smooth(x,y,filter=:GCVSmoothedNSpline, lambda=10.0^1)
+#y_sv = smooth(x,y,method="savgol",window_length=9,polyorder=3)
+#y_gcv = smooth(x,y,method="GCVSmoothedNSpline", ese_y = std(y))
+y_whit = smooth(x,y,method="whittaker", Lambda=10.0^1)
 
-ese_noise = sum((y - y_perfect).^2)
-ese_sg = sum((y_sv-y_perfect).^2)
-ese_gcv = sum((y_gcv-y_perfect).^2)
-ese_whit = sum((y_whit-y_perfect).^2)
+ese_noise = sum((y .- y_perfect).^2)
+#ese_sg = sum((y_sv.-y_perfect).^2)
+#ese_gcv = sum((y_gcv.-y_perfect).^2)
+ese_whit = sum((y_whit.-y_perfect).^2)
 
-@test ese_sg < ese_noise
-@test ese_gcv < ese_noise
+#@test ese_sg < ese_noise
+#@test ese_gcv < ese_noise
 @test ese_whit < ese_noise